// _tools/nextjs-related-files/pages/public-profile/profiles/crm/interfaces.ts
export interface ICRMGeneralInfoItem {
  label: string;
  info: string;
  type?: number;
}
export interface ICRMGeneralInfoItems extends Array<ICRMGeneralInfoItem> {}

export interface ICRMAttributesItem {
  label: string;
  info: string;
}
export interface ICRMAttributesItems extends Array<ICRMAttributesItem> {}

export interface ICRMAttributesProps {
  url: string;
}

export interface ICRMApiCredentialsProps {
  url: string;
}

export interface ICRMDealsItem {
  name: string;
  ammount: string;
  date: number;
  label: string;
  color: string;
}
export interface ICRMDealsItems extends Array<ICRMDealsItem> {}

export interface ICRMActivityProps {
  url: string;
}

export interface ICRMRecentInvoicesItem {
  icon: string;
  number: string;
  date: string;
  ammount: string;
}
export interface ICRMRecentInvoicesItems extends Array<ICRMRecentInvoicesItem> {}